64-bit intermediate to prevent uptime-dependent hue discontinuities
With strip.now in milliseconds, the fast-path overflows after ~7.5 hours and the slow-path after ~54.2 hours, causing visible hue jumps in continuous uptime scenarios.

=> Cast the multiplication operands to uint64_t before division.

+++ usermods/usermod_v2_animartrix/usermod_v2_animartrix.h
@@ -152,8 +152,8 @@ class ANIMartRIXMod:public ANIMartRIX {
 
 		if (cycle_hue) {
 			unsigned tt = strip.now;
-			if (SEGMENT.intensity > 128)      tt = (tt * (31U + SEGMENT.intensity - 127)) / 32U;  // faster up to 4x
-			else if (SEGMENT.intensity < 127) tt = (tt * 22U) / (21U + 127 - SEGMENT.intensity);  // slower down to 1/6
+			if (SEGMENT.intensity > 128)      tt = (uint64_t(tt) * (31U + SEGMENT.intensity - 127)) / 32U;  // faster up to 4x
+			else if (SEGMENT.intensity < 127) tt = (uint64_t(tt) * 22U) / (21U + 127 - SEGMENT.intensity);  // slower down to 1/6
 			hueshift = (tt << 4) | (tt & 0x0F);
 		} else {
 			hueshift = (128 - SEGMENT.intensity) * 256;
